rename scripts module into analysis
move high-lelve interface code into analysis sub-package
