+++ gammapy/analysis/__init__.py
@@ -1,0 +1,3 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Gammapy high-level interface (analysis)."""
+from .analysis import *
