+++ gammapy/analysis/analysis.py
