+++ gammapy/analysis/config/docs.yaml
