+++ gammapy/analysis/config/model.yaml
