+++ gammapy/analysis/config/schema.yaml
