+++ gammapy/analysis/config/template-1d.yaml
