+++ gammapy/analysis/config/template-3d.yaml
