+++ gammapy/analysis/config/template-basic.yaml
