+++ gammapy/scripts/__init__.py
@@ -1,3 +1,2 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Gammapy command line interface (scripts)."""
-from .analysis import *
