+++ setup.cfg
@@ -56,7 +56,7 @@ console_scripts =
     gammapy = gammapy.scripts.main:cli
 
 [options.package_data]
-gammapy.scripts = config/*
+gammapy.analysis = config/*
 gammapy.irf.tests = data/*
 gammapy.modeling.tests = data/*
 gammapy.catalog.tests = data/*
